+++ src/main/java/calculator/view/OutputView.java
@@ -1,10 +1,10 @@
 package calculator.view;
 
 public class OutputView {
-    
+
     private static final String RESULT_OUTPUT_MESSAGE = "결과 : ";
 
-    public void printSumResult(int result) {
+    public void printSumResult(long result) {
         System.out.println(RESULT_OUTPUT_MESSAGE + result);
     }
 }
